Added contacts listing UI
